Commented out special code for 'deploy' goal. It prevents actual deploy
The pom contains code that copies the generated .so file to its
final destination under /usr/lib or /usr/lib64 during the deploy phase.
That approach has some drawbacks:
a) The build then must run under root privileges
b) The actual deploy that copies the artifacts to its remote
   repository was disabled.

We need to figure out a better way to install the .so into /usr/lib.