Update annotation-driven SQL generation to Java 6.
Remove the dependency on org.objectweb.asm in favor of the
annotation processing facilities now built in to javac as of
Java 6. Add annotations to override the auto-determined type
for function parameters and/or supply default values, and to
supply arbitrary SQL commands to be included in the generated
deployment descriptor.