SQL/JRT 2003: Jar may have > 1 deployment descr
Moving the annotation examples into pljava-examples creates a jar
with two deployment descriptors, the old one for the non-annotated
examples and the generated one. That's actually ok per SQL/JRT 2003,
but small changes to Commands.java and a new sqlj.jar_descriptors
table are needed to remove the former arbitary restriction to one
descriptor only.

Eventually the non-annotated examples could all be annotated,
leaving the generated descriptor as the only one. But as the
limit to one descriptor was at variance with the standard anyway,
it seemed worth removing.