Make maven delay copying the service config file
The annotation processor service configuration file can't be
in resources/ because maven copies those resources to the target
directory before compilation, and in a clean build javac will fail
trying to instantiate a processor it hasn't compiled yet.

Moved the config file to a new directory late-added-resources/
with an addition to pom.xml to do the copy only just before producing
the jar.