Get annotated examples working in pljava-examples
For now, nothing done but to dump the annotated examples in with
the other ones (and change the name of one method that duplicated
one of the non-annotated ones), and further annotate the trigger
example to create the tables it refers to. (Also, one example
referred to an example.properties file, so one has been created.)

For now, the result is a jar with two deployment descriptors (which
is supposed to be ok per SQL/JRT 2003, so is now accepted): the
original one and the generated one.

Notes on things I know aren't perfect yet:

1. Maven doesn't put things in the jar in a deterministic order.
The generated ddr depends on the javatest schema and _properties
type created in the old ddr, but sometimes "mvn clean install" will
put the two ddrs in the jar in the other order. This is just a
maven issue; pljava correctly follows the spec and uses the order of
the ddrs found in the jar. It wouldn't be a problem if the two ddrs
had no interdependencies, or if the maven build could be tweaked
to force a certain order in the jar, or if the old ddr were done away
with entirely by annotating all the old examples so the generator
generates everything.

2. The generated ddr only reflects annotations seen in the compile run.
If a non-clean build run only runs javac on a few changed files, the
generated ddr can be incomplete. So for now it's best to run javac on
everything (or do a mvn clean before building). There may be a way to
improve the annotation processor to save a record of all the files that
would need to be reexamined in a partial rebuild. What I don't know yet
is how (or whether) that can be communicated back to javac to get those
files examined in a later round.

3. Naturally until issue #8 is fixed, none of the no-arg example
functions can actually be used.

4. The old examples.ddr contains a set search_path making javatest the
current schema, which causes deployRemove to fail later. The trouble is
that if deployRemove alters the search path on the current schema, it will
try to restore the old value after undeploying, and that fails if the
undeploy made that schema go away. Possibly deployRemove could be made to
just fail silently if that happens, but that's an issue for another day.