Fix obsoleted query in sqlj.replace_jar.
Ken Olson caught this a year ago in his pull request #29, but
it was my oversight; when rearranging the sqlj relations back
in #10 to accommodate more than one deployment descriptor, I
updated the sql in install_jar and remove_jar (and deployInstall
and deployRemove), but I just totally overlooked replace_jar. Oops.