Run deployment descriptors in correct order.
Previously determined the order of multiple deployment descriptors in
a single jar according to the order of those entries as stored in the jar
(used in that order for install, and that order reversed for remove).

But that wasn't correct. I got my hands on 2003 and 2006 drafts of the
SQL/JRT spec and they both clearly say it is the order _as the entries
are listed in the manifest_ that matters (again, in that order for install,
and the reverse for remove).

This should be a welcome improvement, because I had noted back in
commit 0edc9e5f that maven doesn't always put things in a jar in
the same order, and that was causing the pljava-examples jar to be
broken about half the time (for autodeployment anyway). But the manifest
is a static file listing the ddrs in the right order, so as long as
maven doesn't reorder it while putting it in the jar, that behavior
should now be stable.