Workaround the problem reported in #39.
The annotation processor in javac runs in multiple 'rounds', as long as
generated source files keep appearing, and one final round after they
don't. This code used to save some mapping of Class objects to the
javax.lang.model objects (TypeElement/TypeMirror) until the final round,
which worked in Java 6, but as of 7 one gets back different model objects
in different rounds, for the same types, and they don't match, so type
mapping breaks.

This workaround moves all those lookups to before or during round 1, when
a consistent set of model objects can be looked up. So, it will work as
long as all the source files that might contain pljava annotations will be
found in round 1. That should always be the case unless someone is using a
very fancy build with _other_ annotation processors generating new source
files with pljava annotations that have to be processed in additional
rounds. For now, that won't work, because their types won't seem to match
what was computed in round 1. So don't do that.

http://bugs.java.com/bugdatabase/view_bug.do?bug_id=8038455 might refer
to this problem, and promises a fix in Java 9, for what it's worth.