Get maven to build a basic (all defaults) tech-ref site.
With these pom additions, `mvn clean compile site site:stage`
will build a directory `target/staging` that contains a web site
with a bunch of autogenerated technical reference and javadocs
for the project. This should not _substitute_ for the user-oriented
documentation on the wiki, but should be linked to _from_ the wiki
for easy access to deeper detail.

In the default output the javadocs are a bit buried (click a project
module, then click Project Reports), but that should be ok as they can
be linked to easily from the user docs on the wiki and that is probably
how most people would get there. I'm sure there are lots of possible
configuration tweaks in what maven can generate, but for now I think it
would be a big improvement just to have tech refs on the site at all.

I think the `gh-pages` site would be an ideal place for this generated
stuff to be pushed. There is stuff in the `gh-pages` branch right now
but it seems to be the same stuff that's on the wiki, and the wiki is
probably the right place for it. So a clean empty `gh-pages` orphan branch
could be made, and `tada.github.io/pljava` could be where the autogenerated
reference material goes. I believe github offers a dedicated maven plugin
that can automagically push the generated site there.