The rest of "Bugfix 1011181"
This was one problem being solved two ways. First there was code
to double any percent signs in a string that would be passed to elog.
But then myrkraverk in 2012 (622d0ac) got a compiler warning about
using a non-constant format string, and changed the elog call to
pass the supplied string as an argument to `%s` format, which is the
cleaner solution anyway, but didn't remove the old code, so `%` signs
were printing out doubled.