More POM adjustment to tweak the generated site.
Includes some spelling/naming changes for consistency, and sets
relativizeDecorationLinks to false to make it possible to refer to
the main github site as the project URL without messing up links
on the github.io site. (relativizeDecorationLinks was added per
https://issues.apache.org/jira/browse/MSITE-159)

At the moment, intermodule JavaDoc references will be wrong. It's
always something.