Revert to the github.io site for the Project URL.
The assumption that the Project URL *is* the site Maven is generating
seems to be baked too deeply into things like
https://maven.apache.org/plugins/maven-javadoc-plugin/javadoc-no-fork-mojo.html#detectOfflineLinks
to be worth fighting.

So the generated site will just have to have some custom content
to direct people to the wiki and main github site for most purposes.