Enable site publishing with trajano/wagon-git.
This wagon-git plugin was mentioned by others as working better
for multi-module site publishing than the official github site plugin.

The magic happens when someone with push access to the project repo
runs mvn site:deploy. Credentials for the push aren't included among the
project files for obvious reasons ... instead the local settings.xml for
the person running maven must contain a <server> entry with the same ID
(site.pljava.tada.se) used in the pom.xml, and that server entry needs
to have a <privateKey> entry with the name of the SSH identity file to
be used.

To make this work with ssh-agent (instead of embedding the passphrase
in settings.xml, yuck!), the version of trajano/wagon-git must be
at least 2.0.4-SNAPSHOT with
[pull request 14](https://github.com/trajano/wagon-git/pull/14) applied.

One day when that functionality appears in a non-snapshot release, the
oss.sonatype snapshot repo can be removed in pom.xml.