MSVC profile using profile more, properties less.
This will have to be tested by someone with MSVC access. Its behavior
should be close to what I _think_ it should do. Possible nits: it puts
the MSVC extra include paths late in the list; if that's a problem
because they need to be early, I found syntax for that too, only a
little bit uglier. Also, an MSVC link will mention libjvm twice, once
in the right place (I think), but also still in the wrong one; I _think_
that should be harmless. Nothing another barrage of XML can't fix, if
it's a problem.