Squash silly conversion-to-bool warnings.
nar-maven-plugin turns on such aggressive gcc warnings that it ends up
warning about a = b ? c : d where the things are all bool. Sure, strictly
by the standard, c and d get integer-promoted and then have to be
narrowed back to bool, but gcc would know better than to give a warning
if bool really were _Bool. The thing is, PostgreSQL's c.h is defining it
to be char, so gcc doesn't know better.

If it were happening a ton of places I would just live with it or turn off
the warning, but changing two places isn't so bad. Rewrote using 'if' to
look less dorky than hanging a (bool) cast on a perfectly obviously
boolean expression.