Fix varlena UDT breaking if longer than 32767.
Spotted among the flurry of gcc warnings, most of which were nuisances.

Declaring `dataLen` as `int16` makes sense if looking at the `typlen`
catalog attribute, but then in `coerceScalarDatum` it can later get used
for a `varlena` size, which can be up to 1GB-ish.

It wasn't strictly necessary to change it also in `coerceScalarObject`
(`buffer.len` gets used for the var size there), but I did for consistency.
Other `dataLen`s later in the file were already `int32`, so now they all
are.

Closes #52.