Squash long-to-int narrowing warning in Timestamp.
I was baffled why `tm_gmtoff` would ever be declared `long` until I saw
Tom Lane's comment [just this last May][revelation]: "The fact that
`tm_gmtoff` is declared as `long` is no doubt a hangover from days when
`int` might be only 16 bits ...."  I'm not sure 'hangover' is the word
he meant but it seems strangely appropriate.

His solution was to cast to `int` without compunction, and so is mine.
There's even the luxury of saying `int32` and never again worrying about
whether it's big enough for ±86400 seconds.

[revelation]: http://git.postgresql.org/gitweb/?p=postgresql.git;a=commit;h=c90b85e4d9e4ae3bc26459cc54697e1adaa4315f