Now Snippet can carry implementor-name.
Nothing sets it yet, but it now defaults to PostgreSQL for everything, in
recognition of the fact that hardly any of the PG SQL code that gets
generated here will be in the exact ISO-prescribed syntax of the exact
five allowed, unadorned SQL statements.

Yes, that makes DDR files bigger. The zip format will compress them well.