Snippet implicitly requires its implementor-name.
That way, a snippet that will test some condition and enable that
implementor-name at install time can be declared to provide it, and
be emitted earlier. However, not to insist on supplying a snippet
that explicitly provides every implementor name ... some could just
be selected by the user with SET LOCAL pljava.implementors TO ...
before installing. So the implicit requirement will be counted in
indegree initially, to delay such a snippet as far as practical, but
only until a cycle-breaker releases it when no other progress can
be made.