Eliminate threadlock ops in string conversion.
The Java methods related to charset encoding/decoding may be called
repeatedly, and they don't require the threadlock to be released and
reacquired. I don't measure much difference in timing (I don't really
have a good "average text" corpus to test on; the test case for this
bug is worst case because it uses all of the biggest characters.)

Even without a compelling timing difference, the Java charset encoders/
decoders aren't thread safe, so I feel just that much better holding
on to the lock.