Conform to BaseUDT rename, add MappedUDT.
Supporting MappedUDT is much simpler. It emits a
CREATE TYPE foo AS ( ... structure ... )
if a structure is provided, or not, if it isn't, followed by a
SELECT sqlj.add_type_mapping( ... ), and that's it.

Much javadoc also added.

This commit ends with pure renames of the Point and ComplexTuple
examples in preparation to make them annotation examples ... breaking
compilation until they are fixed up in the next commit, but git can
see where they went.