ComplexTuple.logAndReturn needs IMMUTABLE.
That's how it was declared in the hand-maintained examples.ddr, and
letting it default to VOLATILE makes it produce TupleDesc reference
leak warnings for some reason, so I've added IMMUTABLE to the annotation.

At the same time, onNullInput=RETURNS_NULL is clearly needed, just by
glancing at the function, which does no null checking. That's also true
of Point, so that's added both places. This demonstrates an advantage of
the SQL generation ... the place where you put the annotation is exactly
where you're looking when you're looking at the code.