Late minor corrections.
Javadoc for 'implementor' in BaseUDT and MappedUDT showed it had been
copy/pasted from Function. package-info.java did not document new
ddr.implementor. It also turned out that ddr.implementor can't use an
empty string to mean "don't wrap" because javac passes such an option
to the processor as if it were null or not specified ... so
-Addr.implementor=- has been reserved for that purpose.