Resolve OS X build problems (issue 63).
In the nar-maven-plugin, 'shared' and 'plugin' are interchangeable on
platforms other than Darwin, but on Darwin they produce two different
link commands (a 'plugin' gets the '-bundle' option).

The distinction (on Darwin) is a shared object is something exporting
some API that the process loading it will call into, but a plugin/bundle
also expects to be able to call back into API provided by the thing that
loaded it. PL/Java fits in the second category, so may as well call it a
plugin; won't change much on other platforms, but makes a difference on
Darwin.

When linking a plugin, the Darwin linker wants the -bundle_loader option
to identify the program that will be loading the plugin (postgres, in
this case). It verifies that the symbols unresolved in the plugin itself
really will be found in the program that loads it.

Update the packaging script to reflect the change from shared to plugin
in the object path.