Change linker output type shared -> plugin.
In the nar-maven-plugin, 'shared' and 'plugin' are interchangeable on
platforms other than Darwin, but on Darwin they produce two different
link commands (a 'plugin' gets the '-bundle' option).

The distinction (on Darwin) is a shared object is something exporting
some API that the process loading it will call into, but a plugin/bundle
also expects to be able to call back into API provided by the thing that
loaded it. PL/Java fits in the second category, so may as well call it a
plugin; won't change much on other platforms, but might work better on
Darwin.