Ok, simplify PGSQL_*_VER too.
Taking the PKGLIBDIR stuff out of the pom raised the question, what
were PGSQL_{MAJOR,MINOR,PATCH}_VER doing in there as well? The answer,
pretty much nothing: they are used all over the C code, but may as well
be derived in pljava.h from PG_VERSION_NUM. That's the way
PG_MODULE_MAGIC is computed, and our values had better match.

They were used in build.xml, as part of a bewildering check for whether
MSVC is new enough to have the rint() function but the PG version is too
old to know about it in pg_config.h. But that check (the second part,
anyway) can be done more concisely just by looking in pg_config.h.

One day, the pom might again want to refer to the PG version, for
example if it wants to put a PG version classifier on the built
artifact, but by that time, we may have found a tidier way to do it
than with a separate ant build file.