Type_getCoerceOut correct statement order.
Ken Olson pointed this out as a case of 'dead code', but in fact it was
live use of an unassigned variable just before the function call that
was going to assign it. Had to be an editing mixup, has been that way
a very long time, and has probably not caused more problems only because
it's statistically rare for a random stack location to be exactly equal
to InvalidOid.

Scratches the surface of #65. A full solution for that issue will have to
come later, but this much was an obvious Heisenbug with an obvious fix.