Unbreak packaging subproject.
Ken Olson reported: this branch "has the pgsql.properties file in the
wrong place during the packaging step ... Changing the build.xml file in
packaging to reference target finds the file, but then the name is
incorrect because the properties file no longer contains the version
numbers."

That was the one place I overlooked that was using those version numbers
for anything, and it is just to put a -pgm.n- classifier on the generated
zip or targz file, so it's simpler to just compute a pgm.n classifier to
go in the properties file.

But it turned out the name produced by packaging has been incorrect
all along anyway, because it tries to assign ${naraol} by finding the
path of a history.xml file in pljava-so/target, and the nar-maven-plugin
has had a bug maven-nar/nar-maven-plugin/issues/106 that didn't get fixed
until release 3.2.0 of that plugin, where instead of a history.xml file,
there was a history.xmlgiantstringofdigits file, which was not recognized
in the packaging build.xml, so naraol did not get set, and the generated
archive had a literal '${naraol}' in its name....

So, updating to the latest greatest nar-maven-plugin release.