Don't fail in remove_jar if schema dropped.
The pljava-examples jar has long had an annoying behavior because
its remove actions drop its schema, and if deployRemove had changed
the schema's classpath to include the jar, it would fail upon trying
to restore the original classpath on the now-nonexistent schema.

Resolved by telling deployRemove it's ok for the classpath restoration
to fail if the reason is the schema's gone. All references to the removed
jar will be gone from classpath_entry anyway; referential integrity
sees to that.