Give Function_getFunction a simpler contract.
It will now always be true that currentInvocation->function is set
when getFunction successfully returns. Formerly that was true on the
first call when the Function had to be constructed, but would not be
true otherwise, and had to be done later in invoke or invokeTrigger.

Also now retrieve the typeMap only if the function is not a UDT.
It was being stored in func.nonudt and the finalizer would only free it
in the nonUDT case, so setting it in all cases might have been a mistake.