Fix overlooked consequence of pull req 68.
Deriving PGSQL_*_VER from PG_VERSION_NUM in pljava.h changed
behavior in one case where an #if protected an #include
*before* pljava.h was included.  (There's another instance in AclId,
but that is for a pre-8.1 condition that can be removed entirely.)

Note: PG_VERSION_NUM itself appeared in 8.2, so that is now a version
minimum. Given that, it would be even safer (and reduce verbosity) to
dedicate one future commit to just rototill all of the sources and
change PGSQL_*_VER conditionals to direct ones on PG_VERSION_NUM.

Committing without a separate pull request, as follow-on to PR 68.