InstallHelper groundwork() ensures good schema.
The language handler functions are created with CREATE OR REPLACE
so they will be sure to refer to the newly-specified native library,
without requiring a cascading drop of everything depending on them.

Most of what belongs in the schema can be autogenerated into a
pljava.ddr by annotations in Commands.java. No Deployer or install.sql
to keep in sync. (For some reason the byte[] parameters had to be
explicitly SQLType'd bytea, even though the DDRProcessor type mapper
is supposed to know that. A puzzle for another day.)