Add to functions of InstallHelper.
The hello() function is passed information from the backend that
would be nice to be able to refer to in Java properties, and also
constructs an informative message returned to the backend for
display.

Let Backend.c use the JVMOptList only for things that have to
be set there. Anything that's an ordinary system property can
be set in hello.

... and one of those things almost certainly should be
java.awt.headless. In the near-unimaginable other case, it can
be overridden in pljava.vmoptions like anything else.