Factor init sequence into a restartable routine.
The initsequencer routine can pick up where it left off, whether
started by _PG_init, by the internal call handler, or by a GUC hook
if the admin is trying a different config setting.

Continue moving bits out of the initializeJavaVM, etc., routines
into additional stages in the sequencer. Eventually: have a stage where
the JVM is running and a Java method can be called to check for and do
the SQL function declarations, SQLJ schema population, etc.

Ability to re-create the VM to try a new classpath is what I had hoped
for, but in many years of the JNI docs the restriction to one JVM per
process hasn't been relaxed yet, and it seems to mean you can't start
a second even after shutting down the first. So in many cases the admin
will have to exit and start a new session if installation fails at that
point. That's still not too bad.