Use PG 8.3 find_coercion_pathway API.
Ok, this compiles and doesn't break any -examples tests (and will give
errors for unhandled return cases, or a warning when disregarding
domain constraints). The reason it went unnoticed so long is that
Type_getCoerce{In,Out} are really very seldom called. They are almost
dead code--fairly contrived function declarations are needed even to
force them to be called for test purposes.

That seems to be mostly because Type_getCoerce{In,Out} won't be
called if Type_canReplaceType returns true, which it very often does,
because (a) Type_fromOid silently replaces domains with their base
types, and (b) Function builds the method signature by mapping the
SQL return type to a Java type and looking for a matching method, failing
if none is found, rather than looking for a method by its name and
parameter signature only, then validating its return type. So, essentially
by construction, coercion of the return type can't ever turn out to be
necessary.

It could, however, if the AS string gives the Java return type explicitly,
and it requires coercion to the (base type of the) SQL type. SQL/JRT has
no syntax to specify the Java return type, but as an extension it seems
PL/Java does: "Function mapping" in the wiki gives an example where it
precedes class.method, separated by a space. Only the example doesn't
work. That's because the getAS code accepts the syntax only when the
return type is purely alphanumeric (no dots, so non-package-qualified).
Not a recent change, has been that way since 2006. An = is also accepted,
though, even after a package-qualified return type. The example works if
retried with an =.

SQL/JRT, it seems to me, specifies the other approach to method
resolution, that is, find the method by name and parameter signature,
then work out what to do with its return type, but that would be a change
to current PL/Java behavior.

The unimplemented warning for RELABELTYPE to a domain type essentially
can't be triggered, because of the way any domain is replaced by its base
type before consulting canReplaceType. That's harmless for IN parameters.
For the return type, it's a type-safety hole: a PL/Java function can
return a value that isn't valid for its declared result domain. Fixing
that should be possible, but beyond the scope of this issue.

The not-implemented errors for COERCEVIAIO and ARRAYCOERCE can be
triggered, just by constructing exactly the sort of function declarations
where you would expect those things to happen. They would still be
arguably contrived cases, where the AS string specifies a method with
different types than would naturally correspond to the SQL ones.
Nevertheless, the "Function mappping" wiki page says "PL/Java will use
a standard PostgreSQL explicit cast when the SQL type of the parameter
or return value does not correspond..." and that's not completely true
at the moment, as long as some features available in standard PostgreSQL
explicit casts, like array or I/O coercion, aren't yet implemented.
Again, beyond the scope of issue 65. At least those cases now give clear
errors, instead of crashes or Krueger numbers as I just confirmed in a
build without this change, so this does in fact fix a latent bug.

Note that the single-row result set writers used for functions that
return composite or set-of-composite results have their own coercion
logic completely unrelated to this, and implemented in SPIConnection.java.
There are more moving parts here than I had hoped....