Note historical nature of files in docs/.
All of that information is present in more current versions on the
project site and/or wiki. These old files are interesting, but only
if there's no risk of mistaking them for the current documentation.

Handled by adding an index.html that clearly describes them as
historical. It also makes no sense to have pljava-packaging continue
to zip them up in the package archive as if they were really the docs.

Whether it should put anything else their in place of these files is
an open question. Maybe the install process should just COMMENT ON
the sqlj schema and language entries, and mention the online doc URL
there.