PL/Java as a PostgreSQL 9.1+ extension.
Because PL/Java installation is already touched off by a simple
library LOAD (even on pre-9.1 databases without the extension
framework), extension support is largely a matter of having
extension scripts that touch off a LOAD. But the library does need
to detect that case and change some behaviors slightly. When LOAD
is used directly and something fails because the settings aren't
right, Backend reports a WARNING rather than an ERROR, to make it
as easy as possible to keep explore settings and eventually get them
right, without forcing a rollback. However, under CREATE EXTENSION,
it must be reported as an error, or the extension machinery will
think all is well.

Also, playing with the settings after CREATE EXTENSION failed can
result in getting PL/Java successfully installed but not as an
extension--easily remedied with CREATE EXTENSION FROM UNPACKAGED,
so a NOTICE suggesting that is reported in that case.

The method of finding the module pathname from the LOAD argument
doesn't work under CREATE EXTENSION (ActivePortal still refers to
the CREATE EXTENSION command itself, not the LOAD command in the
script), so as a workaround the script creates a temporary table
with the module pathname in it. This turns out to be useful anyway,
because it allows distinguishing the case of PL/Java itself being
installed by CREATE EXTENSION from the case where PL/Java is already
installed and its library is getting loaded in passing during
CREATE EXTENSION for something that depends on it. (Actual support
for PL/Java-managed extensions will have to come later, but the
case needs to be recognized even to say "you can't do that yet.")

Reorganized the archives created by pljava-packaging so files are
at directory prefixes like pljava/pkglibdir, pljava/sharedir, etc.
That makes it easy to know where things should go in a manual
installation, but also opens the possibility of a self-extracting
archive that will use pg_config to look up the corresponding paths
at extraction time.