Make CRLF consistent for text files in zip.
This is in preparation for writing a jar manifest that
declares them as content-type text/plain, which has a
strict CRLF newline style prescribed in RFC 2046.

Then a self-extractor class can be included in the jar
to extract those files using the destination platform's
newline style.

XXX at the moment this only filters the three known-name
extension files, and not any other .sql files that would
be picked up by the *.sql fileset. Probably need a script
ResourceCollection for that.