Update to use Manifest/Attributes, new keys.
Will still honor charset= on content-type, but the Java
powers-that-be didn't go for that on bug 4310708, so there
needs to be a dedicated key to mean the same thing:
_JarX_CharsetInArchive
and there is also _JarX_CharsetWhenUnpacked to override
the platform default encoding for any given entry. Both can have
defaults in the main attributes.

Instead of opening the jar twice, just goes linearly through
a JarInputStream ... this requires the manifest to appear before
any entry to be extracted, but that does not seem too much to ask.
I did verify that it doesn't fail if a separate directory entry
for META-INF/ precedes the manifest.