Workaround Windows creating_extension visibility.
Should now detect (in most cases?) when an extension is
being created, even in versions where creating_extension
isn't visible in Windows. Test depends on seeing the command
in ActivePortal; I am not sure what contexts could be contrived
where that wouldn't work right, but ordinary foreseeable cases
seem to work.

Got rid of pljavaInExtension: the idea that two cases have to be
distinguished (loading PL/Java itself as an extension, or using it
in the creation of some other extension) was sound, but the second
case isn't something that can be checked once at load time; it needs
a backend function that sqlj.install_jar can invoke whenever needed.