Improve message in strange javac error case.
Earlier in the week I saw truly perplexing errors that stemmed
from a simple missing or misspelled import, but javac, instead
of reporting the error, would supply DDRProcessor with an instance
of an internal Error class where the unresolvable value should have
been. So now, at least recognize that Error class and produce a more
helpful message. Now, of course, I am unable to test the change,
because javac (same version! same machine!) is now doing the right
thing every time. At least with this change, if the right conditions
are ever hit again for javac to misbehave as it was earlier, the
message should be more helpful.