Cluster name as property and in jps title.
Make the cluster name easily available if it is set (to a
nonempty string, in PostgreSQL versions where it is available),
as a java system property and in the title visible to jps
and jvisualvm.

In passing, lose the default vmoption to lock out jvisualvm local
connections. Noah Misch persuades me that PostgreSQL itself doesn't
make special efforts to prevent things that already require access
as the postgres user on the server host, and it would be unpostgresy
to do so here. Instead, simply document how to disable attachment,
if the admin so desires.