Catch compiler-caused classcast confusion.
When the compiler hasn't made sense of an annotated element because
of some error in the source, it may pass the annotation processor an
element of an unexpected type. The right response seems
to be to silently return (before anything that risks a class cast
exception), as the compiler will produce its own messages about what
the true problem was.