Check for base UDT if no other type mapping found.
Addresses issue #99. The strategy of registering base UDTs on the
first call to their input/output/send/recv functions breaks down if
the first reference in a session to such a type is one that doesn't
involve calling those functions (for example, an existing table has
a column of that type, and a query wants to pass it to a Java
function accepting that type; PostgreSQL has no reason to think it
should call an I/O function in that case).

This change adds one more step to the rules in Type_fromOid. If no
previous step succeeded, and just before punting to String, take a
close look at the type to see if it could be a Base UDT.

That's a tedious check that involves following the type's input/
output/receive/send Oids via prolang back to their language Oids, and
(one could stop here by assuming the language will always be named
'java' or 'javaU', but if not) from there to the language's call
handler C function and its dynamic library path, which should match
PL/Java's. A future optimization could cache language Oids that are
found to refer to PL/Java, but it might not be needed often enough
to matter.