Lay groundwork for typmod support.
To fully support type modifiers, including the three-argument forms
of input/receive functions, will take more thought than 1.5.0 can wait
for. But to release 1.5.0 with the SQL generator emitting the
one-argument forms would invite headaches later, because CREATE OR
REPLACE FUNCTION can't change the parameter list, and DROP/reCREATE
would cause cascading disruption. So, make sure the generator already
emits the three-argument form, and just be sure to fail if a non-default
typmod is in fact passed, with more complete support to be added later.

By happy chance, it seems that PostgreSQL only very rarely passes non-
default typmods to the input/receive functions--possibly only during
COPY operations. (Those were the only way I was able to test this
change.)

In most other cases where a typmod is used, PostgreSQL relies on a
typmod application cast in a separate step, not on passing the typmod
to input/receive. PL/Java is already usable to implement typmodin,
typmodout, and cast functions, so in limited testing it seems possible
already to do types with typmods, maybe as long as COPY operations on
them are not needed. Therefore, add an example that illustrates it.