Plug composite UDT TupleDesc leak.
In the composite case, UDT was keeping a reference to the type's
TupleDesc, causing PostgreSQL's reference counter to report a leak.
Not a growing, memory leak: only one TupleDesc for each type would
be kept. Of more concern would be how to invalidate it if the
underlying type is altered. More convincing to just grab it from the
type cache whenever needed. More at:

http://www.postgresql.org/message-id/flat/56DF8122.20607@anastigmatix.net