Add Johann Oskarsson's TupleDesc leak test.
This was in the pgFoundry bug report 1010962, a gnarly test of
a composite type holding composite types, then put in an array
to see how many TupleDesc reference leak warnings could result.

The patch by Heikki Linnakangas was described in that report as
"(at least) partial fix". While this test does not use UDTs and
is therefore not strictly related to the UDT fixes just made in
this branch, I definitely see zero TupleDesc reference warnings
from it now, so this is as good a time as ever to get it committed
as a regression test.