Edit Solaris/g++ AOL properties slightly.
The AOL properties can't be the place for hardcoded references
to version-specific PostgreSQL installed locations. Those have to
come from pg_config.

It looks like these are being used to build pljava.so with a RUNPATH.
If it needs one, that will have to be set up in pom.xml (something
that already is needed to accommodate (at least some) EDB community
builds, reported in issue #103).