Allow a link RUNPATH on mvn command line.
The pljava.libjvm_location variable was supposed to make
runpath/ldconfig/LD_LIBRARY_PATH hacks a thing of the past, but there
are still situations (*cough* EnterpriseDB PostgreSQL distributions with
substitute system libraries stowing away in libdir *cough*) where the
old tricks have to be dusted off again. This documents the cases where
they may be needed, and adds a mvn command option for simple cases of
setting a RUNPATH.

Also makes pgtypes/pq/ecpg link inclusion optional.

The build machinery has been including these three libraries
in the link for some time now, but it is not necessary on all
platforms and, in fact, where it isn't necessary, it's better
not to. By not including them, more cases that would otherwise
require runpath tricks can be avoided.

Now they are added to the link only when the -Plinkpglibs option
is used on the mvn command line. If PL/Java builds and works without
it, great. Otherwise, try again, adding that option; if it works,
please report the platform and configuration so we know where it
matters.

It would have been nice to make this Maven profile activeByDefault
and simply encourage brave souls to try deactivating it, but alas,
Maven's activeByDefault doesn't really work that way ... all
activeByDefault profiles are automatically deactivated if *any*
profile is selected on the command line.